+++ src/main/java/dev/railroadide/railroad/project/creation/step/UpdateFabricGradleFilesStep.java
@@ -24,8 +24,8 @@ import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
-public record UpdateGradleFilesStep(FilesService files, HttpService http, TemplateEngineService templateEngine,
-                                    String branch, boolean includeSettingsGradle) implements CreationStep {
+public record UpdateFabricGradleFilesStep(FilesService files, HttpService http, TemplateEngineService templateEngine,
+                                          String branch, boolean includeSettingsGradle) implements CreationStep {
     private static final String TEMPLATE_BUILD_GRADLE_URL = "https://raw.githubusercontent.com/Railroad-Team/Railroad/%s/templates/fabric/%s/template_build.gradle";
     private static final String TEMPLATE_SETTINGS_GRADLE_URL = "https://raw.githubusercontent.com/Railroad-Team/Railroad/%s/templates/fabric/%s/template_settings.gradle";
